Move MUI Theme to styles Directory
