+++ styles/theme.ts
@@ -1,0 +1,13 @@
+import { createMuiTheme } from '@material-ui/core/styles'
+
+const theme = createMuiTheme({
+  overrides: {
+    MuiAppBar: {
+      colorPrimary: {
+        backgroundColor: '#194661',
+      },
+    },
+  },
+})
+
+export default theme
